CDSK-817 - move regex tests to integration/test_branch_regexes
